backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic==2.9.1
pydantic-settings==2.6.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
itsdangerous==2.2.0
google-generativeai==0.8.2

